* compile with Wall -Wpedantic with nearly no warning * formating fixes
git-svn-id: svn+ssh://c4science.ch/source/amie/trunk@2918 64fb8e13-074e-4207-b257-2d34fb7690c6

+++ examples/main_def.cpp
@@ -41,7 +41,7 @@ void step(FeatureTree * featureTree )
     int nsteps = 1;
     featureTree->setDeltaTime(0.0004);
     featureTree->setMaxIterationsPerStep(128) ;
-    for(size_t i = 0 ; i < nsteps ; i++)
+    for(int i = 0 ; i < nsteps ; i++)
     {
         featureTree->step() ;
         featureTree->printReport(i == 0);
